requirements.txt
feedparser==6.0.11
psycopg2-binary==2.9.9
